fix integer overflow in scan command entry index display
Cast offset to long before adding displayed count to prevent integer
overflow when scanning with large offsets.

+++ src/main/java/net/consensys/bric/commands/ScanCommand.java
@@ -120,7 +120,7 @@ public class ScanCommand implements Command {
             String keyHex = Bytes.wrap(key).toHexString();
 
             System.out.printf("[%d] Key (%d bytes): %s%n",
-                finalOffset + displayed[0], key.length, truncateHex(keyHex, 80));
+                (long) finalOffset + displayed[0], key.length, truncateHex(keyHex, 80));
             System.out.printf("    Value (%d bytes): %s%n",
                 value.length, truncateHex(Bytes.wrap(value).toHexString(), 80));
         });
